% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{bs}
\alias{bs}
\title{Black-Scholes Function returning euro option prices}
\usage{
bs(S, K, T, r, sig, type = "C")
}
\arguments{
\item{S, }{Spot security price}

\item{K, }{Strike price}

\item{T, }{Time to maturity, fractional years}

\item{r, }{Risk-free interest rate}

\item{sig}{Sigma, option yearly volatility}

\item{type}{Option type "C" or "P"}
}
\value{
option value scalar
}
\description{
Black-Scholes Function returning euro option prices
}
