% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{bs_iv}
\alias{bs_iv}
\title{Numerical computation of options implied volatility using BS}
\usage{
bs_iv(S, K, T, r, market, type)
}
\arguments{
\item{S, }{Spot security price}

\item{K, }{Strike price}

\item{T, }{Time to maturity, fractional years}

\item{r, }{Risk-free interest rate}

\item{market}{Market price}

\item{type}{Option type "C" or "P"}
}
\value{
Option implied yearly volatility
}
\description{
Numerical computation of options implied volatility using BS
}
