% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_bs_cpp}
\alias{calc_bs_cpp}
\title{Black-Scholes Function returning euro option prices
Supports vectors in Rcpp with vectorized calculation}
\usage{
calc_bs_cpp(S, K, T, R, sig, type)
}
\arguments{
\item{S, }{Spot security price}

\item{K, }{Strike price}

\item{T, }{Time to maturity, fractional years}

\item{sig}{Sigma, option yearly volatility}

\item{type}{Option type "C" or "P"}

\item{r, }{Risk-free interest rate}
}
\value{
option value scalar
}
\description{
Black-Scholes Function returning euro option prices
Supports vectors in Rcpp with vectorized calculation
}
