% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_bs_iv_cpp}
\alias{calc_bs_iv_cpp}
\title{Numerical computation of options implied volatility using BS
Supports vectors}
\usage{
calc_bs_iv_cpp(S, K, T, r, market, op_type)
}
\arguments{
\item{S, }{Spot security price}

\item{K, }{Strike price}

\item{T, }{Time to maturity, fractional years}

\item{r, }{Risk-free interest rate}

\item{market}{Market price}

\item{type}{Option type "C" or "P"}
}
\value{
Option implied yearly volatility
}
\description{
Numerical computation of options implied volatility using BS
Supports vectors
}
