% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_rolling_percentile}
\alias{calc_rolling_percentile}
\title{Rolling percentile}
\usage{
calc_rolling_percentile(x, n, na_rm = TRUE)
}
\arguments{
\item{x, }{vector of feature}

\item{n, }{window size}

\item{na_rm, }{to remove NA's or not}
}
\description{
Rolling percentile
}
