% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred.R
\name{fred_get_all_us_rate_tickers}
\alias{fred_get_all_us_rate_tickers}
\title{Gets all US treasury rate tickers
Returns c('DGS1MO','DGS3MO','DGS6MO','DGS1','DGS2','DGS3','DGS5','DGS7','DGS10','DGS20','DGS30')}
\usage{
fred_get_all_us_rate_tickers()
}
\value{
Vector of tickers from FRED
}
\description{
Gets all US treasury rate tickers
Returns c('DGS1MO','DGS3MO','DGS6MO','DGS1','DGS2','DGS3','DGS5','DGS7','DGS10','DGS20','DGS30')
}
