% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred.R
\name{fred_get_rate}
\alias{fred_get_rate}
\title{Load time series data from FRED}
\usage{
fred_get_rate(ticker)
}
\arguments{
\item{ticker}{Name of ticker from FRED}
}
\value{
data.frame
}
\description{
Load time series data from FRED
}
