% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred.R
\name{fred_get_rate_spreads}
\alias{fred_get_rate_spreads}
\title{Gets all possible spreads between tickers
Returns dataframe of spreads}
\usage{
fred_get_rate_spreads(rates, tickers_ordered)
}
\value{
Dataframe of spreads
}
\description{
Gets all possible spreads between tickers
Returns dataframe of spreads
}
