% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred.R
\name{fred_set_key}
\alias{fred_set_key}
\title{Sets FRED API Key from environment variable, fred_key}
\usage{
fred_set_key()
}
\value{
data.frame
}
\description{
Sets FRED API Key from environment variable, fred_key
}
