% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mom.R
\name{get_absr}
\alias{get_absr}
\title{Absolute range from {-1,1} over period
2*((x - minr) / (maxr-minr)) - 1}
\usage{
get_absr(x, n = 20)
}
\arguments{
\item{x}{price series}

\item{n}{lag}
}
\value{
vector
}
\description{
Absolute range from {-1,1} over period
2*((x - minr) / (maxr-minr)) - 1
}
