% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred.R
\name{get_all_rates}
\alias{get_all_rates}
\title{Load time series data from FRED
Returns c('DGS1MO','DGS3MO','DGS6MO','DGS1','DGS2','DGS3','DGS5','DGS7','DGS10','DGS20','DGS30')}
\usage{
get_all_rates()
}
\arguments{
\item{tickers}{Vector of tickers from FRED}
}
\value{
data.frame
}
\description{
Load time series data from FRED
Returns c('DGS1MO','DGS3MO','DGS6MO','DGS1','DGS2','DGS3','DGS5','DGS7','DGS10','DGS20','DGS30')
}
