% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mom.R
\name{get_mom}
\alias{get_mom}
\title{Time series momentum
(x / dplyr::lag(x,n=n))-1}
\usage{
get_mom(x, n = 1)
}
\arguments{
\item{x}{price series}

\item{n}{lag}
}
\value{
vector
}
\description{
Time series momentum
(x / dplyr::lag(x,n=n))-1
}
