% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mom.R
\name{get_psma}
\alias{get_psma}
\title{SMA ratio
RcppRoll::roll_meanr(x, n = n) / RcppRoll::roll_meanr(x, n = k) - 1}
\usage{
get_psma(x, n = 3, k = 20)
}
\arguments{
\item{x}{price series}

\item{n}{short lag}

\item{k}{long lag}
}
\value{
vector
}
\description{
SMA ratio
RcppRoll::roll_meanr(x, n = n) / RcppRoll::roll_meanr(x, n = k) - 1
}
