% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mom.R
\name{get_sma}
\alias{get_sma}
\title{Price to simple moving average
x / RcppRoll:roll_meanr(x, n = n) - 1}
\usage{
get_sma(x, n = 20)
}
\arguments{
\item{x}{price series}

\item{n}{lag}
}
\value{
vector
}
\description{
Price to simple moving average
x / RcppRoll:roll_meanr(x, n = n) - 1
}
