% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{get_tickers}
\alias{get_tickers}
\title{Load time series data from mulitple tickers}
\usage{
get_tickers(tickers, source = "EODHistorical", interval = "1D")
}
\arguments{
\item{source}{Data source (EODHistorical)}

\item{interval}{Interval of data to load (1m,1D,EOD)}

\item{ticker}{Vector of tickers}
}
\value{
data.frame
}
\description{
Load time series data from mulitple tickers
}
