% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markets.R
\name{get_turn_month_days}
\alias{get_turn_month_days}
\title{Get NYSE turn of month trading days indicator}
\usage{
get_turn_month_days(from, to, ndays)
}
\arguments{
\item{from}{Start date}

\item{to}{End Date}

\item{ndays}{Count down/up days}
}
\value{
data.frame
}
\description{
Get NYSE turn of month trading days indicator
}
