% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{pp_boxcox}
\alias{pp_boxcox}
\title{BoxCox transformation, (x^lambda-1)/lambda}
\usage{
pp_boxcox(x, lambda)
}
\arguments{
\item{x}{data}
}
\value{
transformed data
}
\description{
BoxCox transformation, (x^lambda-1)/lambda
}
