% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{pp_boxcox_opt}
\alias{pp_boxcox_opt}
\title{BoxCox transformation w/ find optimal lambda, (x^lambda-1)/lambda}
\usage{
pp_boxcox_opt(x)
}
\arguments{
\item{x}{data}
}
\value{
transformed data
}
\description{
BoxCox transformation w/ find optimal lambda, (x^lambda-1)/lambda
}
