% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{pp_center_scale}
\alias{pp_center_scale}
\title{Center and scale data, x-mean(x)}
\usage{
pp_center_scale(x)
}
\arguments{
\item{x}{data}
}
\value{
centered and scaled data
}
\description{
Center and scale data, x-mean(x)
}
