% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred.R
\name{set_fredr_key}
\alias{set_fredr_key}
\title{Sets FRED API Key from environment variable, fred_key}
\usage{
set_fredr_key()
}
\value{
data.frame
}
\description{
Sets FRED API Key from environment variable, fred_key
}
